refactor: 로또 당첨 카운트, 수익률 계산 관련 클래스 분리
